% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athyrta-package.R
\docType{package}
\name{athyrta-package}
\alias{athyrta}
\alias{athyrta-package}
\title{athyrta: Data Downloading and Caching}
\description{
Tool for downloading and caching data from web apps or other remote sources with a compatible API.
}
\keyword{internal}
