% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athyrta.R
\name{test}
\alias{test}
\title{Connection proof of concept}
\usage{
test(base_url, table, auth)
}
\arguments{
\item{base_url}{Url to a web app or service}

\item{table}{Data table name}

\item{auth}{Unused}
}
\value{
Data frame
}
\description{
Download data from a table
}
\examples{
\dontrun{
  # This example will not work
  result = test("example.com", "data_table")
}


}
